escolhify-demo/requirements.txt
chainlit>=1.0.0
python-dotenv>=1.0.0
aiofiles==24.1.0
aiohttp==3.10.11
annotated-types==0.7.0
anyio==4.11.0
asyncer==0.0.7
bidict==0.23.1
certifi==2025.10.5
chainlit==2.3.0
charset-normalizer==3.4.4
chevron==0.14.0
click==8.1.8
dataclasses-json==0.6.7
exceptiongroup==1.3.0
fastapi==0.115.14
filetype==1.2.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy>=1.21.0
pandas>=1.3.0
pymongo>=4.0.0
scikit-learn>=1.0.0
importlib_metadata==8.7.0
Lazify==0.4.0
literalai==0.1.103
marshmallow==3.26.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
protobuf==6.33.0
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.18
python-socketio==5.14.3
requests==2.32.5
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.3
syncer==2.0.3
tomli==2.3.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uptrace==1.37.0
urllib3==2.5.0
uv==0.9.7
uvicorn==0.38.0
watchfiles==0.20.0
wrapt==1.17.3
wsproto==1.2.0
zipp==3.23.0

jinja2>=3.0.0
aiohttp>=3.8.0
rich>=13.0.0
openai